+++ src/battle_util.c
@@ -10677,7 +10677,7 @@ u32 GetBattlerVolatile(u32 battler, enum Volatile _volatile)
 
 // Sets the value of a volatile status flag for a certain battler
 // Primarily used for the debug menu and scripts. Outside of it explicit references are preferred
-void SetMonVolatile(u32 battler, enum Volatile _volatile, u32 newValue)
+void SetMonVolatile(u64 battler, enum Volatile _volatile, u32 newValue)
 {
     switch (_volatile)
     {
